% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aKIDS.R
\name{aKIDS}
\alias{aKIDS}
\title{alpha-controlled Kernel-based Independence Dual Screening (alpha-KIDS)}
\usage{
aKIDS(x, y, delta, n1.ratio, d, ir, modelx = T, offset = 1, alpha = 0.1)
}
\arguments{
\item{x}{A data matrix where each row corresponds to an observation and each column corresponds to a feature.}

\item{y}{A numeric vector representing the observed survival times.}

\item{delta}{A binary vector of the same length as \code{y}, indicating the censoring status (1 = event occurred, 0 = censored).}

\item{n1.ratio}{A numeric value representing the proportion of observations to be used for KIDS screening.}

\item{d}{An integer representing the number of features to be selected in the KIDS step.}

\item{ir}{A logical value indicating whether to apply inverse regression for estimating the conditional measure. Set to \code{TRUE} to enable inverse regression, or \code{FALSE} to disable it.}

\item{modelx}{A logical value indicating the type of knockoff features to generate. If \code{TRUE}, the function generates approximate second-order modelX knockoff features; if \code{FALSE}, it generates fixed knockoff features.}

\item{offset}{A numeric value representing the offset to be applied in the FDR threshold. The default value is 1.}

\item{alpha}{A numeric vector specifying one or more false discovery rate (FDR) control levels.}
}
\value{
The function returns:
\describe{
\item{ix1}{A vector of indices corresponding to the features selected by the KIDS screening step.}
\item{alpha}{The vector of FDR control levels that were applied during the selection process.}
\item{ix}{A list of vectors, where each vector contains the indices of features selected by the alpha-KIDS procedure with FDR control for each value of \code{alpha}.}
}
}
\description{
This function performs feature screening with FDR control on ultrahigh dimensional right-censored data.
using the alpha-controlled Kernel-based Independence Dual Screening (alpha-KIDS) procedure
The alpha-KIDS method is a two-stage procedure that initially selects a set of potentially important features using the KIDS method,
followed by identifying a refined set under FDR control through a unified knockoﬀ procedure.
}
\examples{
### Simulate data from AFT model
n <- 2000
p <- 5000
rho <- 0.3
Sigma <- rho^as.matrix(dist(1:p,diag = T,upper = T))
x <- rnorm_Choleski(n, rep(0,p), Sigma)
t <- exp(x[,1]+x[,2]+1.5*x[,3]^2+x[,4]+x[,5]/log(1+abs(x[,5]))+x[,6]+1.5*sin(1/2/x[,7])+x[,8]+x[,9]+x[,10]+rnorm(n)/2)
c <- rexp(n,1/7/exp(x[,1]+x[,2]-x[,10]))
y <- pmin(t,c)
delta <- as.numeric(t<c)
### Perform alpha-KIDS
ix <- aKIDS(x, y, delta, n1.ratio=0.25, d=100, ir=F, alpha=c(0.1,0.2))
ix

}
