% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KIDS.R
\name{KIDS}
\alias{KIDS}
\title{Kernel-based Independence Dual Screening (KIDS)}
\usage{
KIDS(x, y, delta, d, ir)
}
\arguments{
\item{x}{A data matrix where each row corresponds to an observation and each column corresponds to a feature.}

\item{y}{A numeric vector representing the observed survival times.}

\item{delta}{A binary vector of the same length as \code{y}, indicating the censoring status (1 = event occurred, 0 = censored).}

\item{d}{An integer representing the number of features to be selected. If not supplied (NULL), the function returns utility measures for all features instead of selected features.}

\item{ir}{A logical value indicating whether to apply inverse regression for estimating the conditional measure. Set to \code{TRUE} to enable inverse regression, or \code{FALSE} to disable it.}
}
\value{
If \code{d} is provided, the function returns a vector of indices corresponding to the \code{d} selected features.
If \code{d} is not provided, the function returns utility measures for each feature, which can be used for feature ranking or selection.
}
\description{
This function performs feature screening on ultrahigh dimensional right-censored data
using the Kernel-based Independence Dual Screening (KIDS) procedure.
The KIDS method selects important features by evaluating two kernel-based R-squared measures:
(1) a marginal measure: kernel-based R-squared between the feature and the censoring status, and
(2) a conditional measure: kernel-based partial R-squared between the feature and the observed survival time, adjusted for the censoring status.
}
\examples{
### Simulate data from AFT model
n <- 200
p <- 5000
rho <- 0.5
Sigma <- rho^as.matrix(dist(1:p,diag = T,upper = T))
x <- rnorm_Choleski(n, rep(0,p), Sigma)
t <- exp(x[,1]+x[,2]+1.5*x[,7]^2+rnorm(n))
c <- rexp(n,1/5/exp(x[,1]))
y <- pmin(t,c)
delta <- as.numeric(t<c)

### perform KIDS
ix = KIDS(x, y, delta, d = floor(n/log(n)), ir = F)
ix

}
